+++ transformers/examples/token-classification/run_ner_synthetic.py
@@ -185,7 +185,7 @@ class DataTrainingArguments:
         metadata={
             "help": "When used with one-to-one mapping, also changes the [CLS] and [SEP] token. Does not change the PAD token."
         },
-    )    
+    )
 
     def __post_init__(self):
         if self.dataset_name is None and self.train_file is None and self.validation_file is None: